ASH stores numbers as longs or doubles. Make it parse numeric input into those data types, rather than into ints or floats with subsequent coercion.
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@11076 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/textui/Parser.java
@@ -2730,7 +2730,7 @@ public class Parser
 			}
 
 			this.readToken(); // integer
-			return new Value( sign * StringUtilities.parseFloat( "0." + fraction ) );
+			return new Value( sign * StringUtilities.parseDouble( "0." + fraction ) );
 		}
 
 		String integer = this.currentToken();
@@ -2746,16 +2746,16 @@ public class Parser
 			String fraction = this.nextToken();
 			if ( !this.readIntegerToken( fraction ) )
 			{
-				return new Value( sign * StringUtilities.parseInt( integer ) );
+				return new Value( sign * StringUtilities.parseLong( integer ) );
 			}
 
 			this.readToken(); // .
 			this.readToken(); // fraction
 
-			return new Value( sign * StringUtilities.parseFloat( integer + "." + fraction ) );
+			return new Value( sign * StringUtilities.parseDouble( integer + "." + fraction ) );
 		}
 
-		return new Value( sign * StringUtilities.parseInt( integer ) );
+		return new Value( sign * StringUtilities.parseLong( integer ) );
 	}
 
 	private boolean readIntegerToken( final String token )
